+++ console2/src/components/organisms/index.ts
@@ -36,7 +36,6 @@ export { default as NewSecretActivity } from './NewSecretActivity';
 export { default as NewTeamActivity } from './NewTeamActivity';
 export { default as OrganizationActivity } from './OrganizationActivity';
 export { default as OrganizationList } from './OrganizationList';
-export { default as OrganizationListActivity } from './OrganizationListActivity';
 export { default as OrganizationOwnerChangeActivity } from './OrganizationOwnerChangeActivity';
 export { default as ProcessActivity } from './ProcessActivity';
 export { default as ProcessChildrenActivity } from './ProcessChildrenActivity';